Added 3 screens embedded inside a NavigationController:
Main Screen
Sign In Screen
Sign Up Screen

Added UI for all the 3 screens. Validation and implementation coming up in next commit.

Home Screen has an ImageView and 2 buttons for Sign In and Sign Up.
When user clicks a button, they are redirected to the respective screen for sign in or sign up.

User can also navigate back to the home screen incase they do not wish to do anything.